Added safeTransferFromWithFees
In order to handle tokens that have transfer fees paid in tokens (in
addition to gas fees), safeTransferFromWithFees has been added to the
CompatibleERC20 library. It checks the balance of the recipient before
and after the transfer and returns it.

compliantTransfer, compliantTransferFrom and compliantApprove have been
renamed to their `safe` equivalent to indicate that they now check the
return of the call and revert immediately if the call fails.